requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-docx==1.1.0
qrcode==7.4.2
pillow==9.5.0
python-multipart==0.0.6
pydantic==2.5.0
